requirements.txt
apache-airflow
boto3
pandas
pytest
sqlalchemy
